+++ src/test/resources/fixtures/sorting/vue/script-setup/expected.vue
@@ -1,0 +1,4 @@
+<script setup lang="ts">
+const button = cn("flex bg-red-500 p-4 text-center font-bold")
+const card = tw`flex bg-red-500 p-4 text-center font-bold`
+</script>