+++ src/test/resources/fixtures/sorting/vue/script-setup/input.vue
@@ -1,0 +1,4 @@
+<script setup lang="ts">
+const button = cn("text-center p-4 flex bg-red-500 font-bold")
+const card = tw`text-center p-4 flex bg-red-500 font-bold`
+</script>