+++ src/test/resources/fixtures/sorting/vue/style-apply/expected.vue
@@ -1,0 +1,9 @@
+<template>
+  <button class="flex bg-red-500 p-4 text-center font-bold"></button>
+</template>
+
+<style scoped>
+.button {
+  @apply flex bg-red-500 p-4 text-center font-bold;
+}
+</style>