+++ src/test/resources/fixtures/sorting/vue/style-apply/input.vue
@@ -1,0 +1,9 @@
+<template>
+  <button class="text-center p-4 flex bg-red-500 font-bold"></button>
+</template>
+
+<style scoped>
+.button {
+  @apply text-center p-4 flex bg-red-500 font-bold;
+}
+</style>